# swagger.yaml
...
swagger: "2.0"
info:
  description: API criada como projeto para o programa Python da Codenation.
  version: 1.0.0
  title: API para armazenamento de Logs de Erros
  termsOfService: http://swagger.io/terms/
  contact:
    email: junior_faccine@hotmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: lawrence-project.herokuapp.com
basePath: /api
...
paths:
...
  /users:
    get:
      tags:
      - users
      summary: Retorna todos os users cadastrados
      operationId: getAllUsers
      consumes:
      - application/json
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        "400":
          description: Nenhum dado cadastrado
      security:
      - token: []
...
  /register:
    post:
      tags:
      - register
      summary: Registra um novo user
      operationId: registerUser
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Objeto User com os dados para serem registrados
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        "201":
          description: User criado
        "400":
          description: Dados invalidos
...
securityDefinitions:
  token:
    type: apiKey
    name: token
    in: header
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      email:
        type: string
      name:
        type: integer
      password:
        type: string
...